\name{rcppbind.demo}
\alias{rcppbind.demo}
\title{R/C++ interface bind demonstration}
\description{
  \code{rcppbind.demo} illustrates several examples how to
  use the \code{rcppbind} R/C++ interface class library.
}
\usage{
rcppbind.demo()
}
\details{
  This function is to provide several examples in using
  rcppbind template class. You can choose which demo to show,
  the R output result will be shown along with both R and C++
  source codes.
}
\author{Gang Liang (gumpleon@gmail.com)}

\references{
  \emph{Writing R Extensions}, available at \url{http:www.r-project.org}.
}

\keyword{misc}
