\name{rcppbind.path}
\alias{rcppbind.path}
\title{Path of R/C++ interface bind header files}
\description{
  \code{rcppbind.path()} gives the path where the rcppbind
  template class files are installed. The only header file
  needs to be included is \code{rcppbind.h}.

  Please read the PDF document and examples of the package
  for more detailed information regarding the usages of the
  template class libray.
}
\usage{
rcppbind.path()
}
\author{Gang Liang (gumpleon@gmail.com)}

\references{
  \emph{Writing R Extensions}, available at \url{http:www.r-project.org}.
}
\keyword{misc}
